requirements.txt
streamlit
requests
huggingface_hub
PyPDF2
python-docx
